Move `Model` trait to models module
For consistency with how the other phases of encoding are organized
(normalization, pre-tokenization), move the `Model` trait to the module
containing models.
